WIP: fix: debug client.log does not close the file
The debug client.log is 1:1 with a soap.Client, and the file is never closed. Unless debug.Flush is closed,
but that would close all open files used for debug for all clients.
Remove debug.Provider.Flush, we don't need this as all debug related files are now closed when a RoundTrip is done.

TODO: sync writes to the client.log, as currently multiple go routines using the same Client results in interleaved log messages.